requirements.txt
firebase-admin==5.0.0
requests==2.25.1
google-cloud-firestore==2.1.2
python-dateutil==2.8.1